requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2022.1
Django==4.2.5
django-crispy-forms==2.0
flatbuffers==24.3.25
frozenlist==1.5.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
keras==2.11.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
Pillow==10.0.1
propcache==0.2.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sqlparse==0.5.2
stripe==11.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
